refactor: refactor inline logo 108
