--- src/components/about/AboutInlineLogo108.tsx
@@ -1,46 +1,0 @@
-const logoWrapClass = "inline text-base-content align-baseline";
-
-function InlineLogo108() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 1024 605"
-      className="inline h-[1.1em] w-auto align-[-0.15em]"
-    >
-      <path
-        d="M26 581c9.75-100.9125 17.0625-188.175 25.8375-292.9875 17.0625-230.5875 113.1-242.775 165.75-253.5 73.125-11.2125 160.8075 5.1825 194.5125 10.2375 100.425 13.65 186.225 103.8375 173.0625 260.325-18.0375 217.425-109.6875 284.2125-218.4 277.3875-124.8-10.725-183.3-138.9375-169.1625-310.5375 16.0875-190.125 118.95-220.35 214.5-227.175 309.5625-27.3 389.5125-9.75 438.75 0 73.6125 15.6 116.5125 58.9875 111.6375 129.1875-5.85 62.4-42.4125 122.3625-132.1125 114.5625-68.25-6.3375-150.15 43.3875-156.975 135.0375-5.3625 64.8375 35.5875 150.15 132.6 157.4625 77.5125 6.3375 148.6875-29.7375 156.975-133.0875 7.8-96.525-49.725-152.1-132.6-159.4125-56.55-4.3875-117.975-52.65-110.175-133.0875 5.3625-69.7125 59.475-106.275 130.65-110.6625 68.7375-2.925 105.3-5.3625 146.965-23.185"
-        stroke="currentColor"
-        strokeWidth="40"
-        fill="none"
-        strokeLinecap="round"
-      />
-    </svg>
-  );
-}
-
-export function AboutInlineLogo108() {
-  return (
-    <span className={logoWrapClass}>
-      <InlineLogo108 />
-    </span>
-  );
-}
-
-export function AboutLogoPhrase({
-  before,
-  after,
-}: {
-  before?: string;
-  after?: string;
-}) {
-  return (
-    <span className="inline-flex items-baseline whitespace-nowrap">
-      {before}
-      {before ? "\u00a0" : null}
-      <span className={logoWrapClass}>
-        <InlineLogo108 />
-      </span>
-      {after ? `\u00a0${after}` : null}
-    </span>
-  );
-}
